Moved fixtures to conftest, added topology marker
